Deprecate implicit component discovery in auto_cli, explicit is better than implicit
